Added SortState and SortCondition files plus initial code.
