+++ lib/axlsx/workbook/worksheet/auto_filter/sort_condition.rb
